message-dialog: deprecate the image field
We had already set the image to be hidden in the .ui file.
This patch removes the image altogether, and deprecates the
property, setter and getter.

If an image is explicitly put with the setter, it is still
shown, so to not break existing users of this API.

Based on a patch by Jon McCann.