GtkPopover: Use display-hint for menus
When an 'iconic' display-hint is specified on a section,
render the actions as a row of iconic model buttons, instead
of like menu items. We're using the verb-icon attribute
here, to keep noun and verb icons cleanly separated.

This works with regular actions, as well as with toggle and
radio actions.